Incorporate config from alpha patch
Applied by running:

    patch -p1 < <(git show e1b403e config.def.h) config.h

e1b403e refers to the local commit that introduces the center patch.

The SHA will go out of date as the local branches that I'm using to
manage my st patches get rebased, but I'm showing the command because
it's a good example of how to keep config.h up to date when new patches
are applied.

What's happening? `git show` shows the change that a patch introduced to
config.def.h (the default config). The `patch` command then applies that
change to config.h.